youtube_upload/requirements.txt
google-api-python-client>=2.127.0
google-auth>=2.33.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
httplib2>=0.22.0
